refactor: extract packet processors into dedicated module
Refactored packet processing logic from aprs_client.rs into a new
packet_processors module with focused, single-responsibility files:

- packet_processors/aircraft.rs: AircraftPositionProcessor
- packet_processors/receiver_position.rs: ReceiverPositionProcessor
- packet_processors/receiver_status.rs: ReceiverStatusProcessor
- packet_processors/server_status.rs: ServerStatusProcessor
- packet_processors/position.rs: PositionPacketProcessor (routes by source type)
- packet_processors/router.rs: PacketRouter (routes by packet type)
- packet_processors/mod.rs: Module declarations and re-exports

Benefits:
- Reduced aprs_client.rs from ~1484 lines to ~865 lines
- Better separation of concerns
- Easier to maintain and test individual processors
- Clearer file organization
