+++ src/aprs_client.rs
@@ -1,13 +1,5 @@
-use crate::fix_processor::FixProcessor;
-use crate::flight_tracker::FlightTracker;
-use crate::receiver_repo::ReceiverRepository;
-use crate::receiver_status_repo::ReceiverStatusRepository;
-use crate::receiver_statuses::NewReceiverStatus;
-use crate::server_messages::ServerMessage;
-use crate::server_messages_repo::ServerMessagesRepository;
+use crate::packet_processors::PacketRouter;
 use anyhow::Result;
-use chrono::{DateTime, NaiveDateTime, Utc};
-use num_traits::AsPrimitive;
 
 /// Result type for connection attempts
 enum ConnectionResult {
@@ -18,8 +10,7 @@ enum ConnectionResult {
     /// Connection was established but failed during operation
     OperationFailed(anyhow::Error),
 }
-use metrics::counter;
-use ogn_parser::{AprsData, AprsPacket, PositionSourceType};
+use ogn_parser::{AprsData, AprsPacket};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::fs::OpenOptions;
@@ -28,7 +19,7 @@ use tokio::net::TcpStream;
 use tokio::sync::mpsc;
 use tokio::time::{sleep, timeout};
 use tracing::trace;
-use tracing::{debug, error, info, warn};
+use tracing::{error, info, warn};
 
 /// Trait for processing APRS packets
 /// Implementors can define custom logic for handling received APRS packets
@@ -718,619 +709,10 @@ impl ArchiveService {
     }
 }
 
-/// PacketRouter implements PacketProcessor and routes packets to appropriate specialized processors
-/// This is the main router that the AprsClient should use
-pub struct PacketRouter {
-    /// Optional archive service for message archival
-    archive_service: Option<ArchiveService>,
-    /// Position packet processor for handling position data
-    position_processor: Option<PositionPacketProcessor>,
-    /// Receiver status processor for handling status data from receivers
-    receiver_status_processor: Option<ReceiverStatusProcessor>,
-    /// Server status processor for handling server comment messages
-    server_status_processor: Option<ServerStatusProcessor>,
-}
-
-impl Default for PacketRouter {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl PacketRouter {
-    /// Create a new PacketRouter without archival
-    pub fn new() -> Self {
-        Self {
-            archive_service: None,
-            position_processor: None,
-            receiver_status_processor: None,
-            server_status_processor: None,
-        }
-    }
-
-    /// Create a new PacketRouter with archival enabled
-    pub async fn with_archive(base_dir: String) -> Result<Self> {
-        let archive_service = ArchiveService::new(base_dir).await?;
-        Ok(Self {
-            archive_service: Some(archive_service),
-            position_processor: None,
-            receiver_status_processor: None,
-            server_status_processor: None,
-        })
-    }
-
-    /// Add a position processor to the router
-    pub fn with_position_processor(mut self, processor: PositionPacketProcessor) -> Self {
-        self.position_processor = Some(processor);
-        self
-    }
-
-    /// Add a receiver status processor to the router
-    pub fn with_receiver_status_processor(mut self, processor: ReceiverStatusProcessor) -> Self {
-        self.receiver_status_processor = Some(processor);
-        self
-    }
-
-    /// Add a server status processor to the router
-    pub fn with_server_status_processor(mut self, processor: ServerStatusProcessor) -> Self {
-        self.server_status_processor = Some(processor);
-        self
-    }
-}
-
-impl PacketRouter {
-    /// Process a server message (line starting with #)
-    pub async fn process_server_message(&self, raw_message: &str) {
-        if let Some(server_proc) = &self.server_status_processor {
-            trace!("PacketRouter processing server message with ServerStatusProcessor");
-            server_proc.process_server_message(raw_message).await;
-        } else {
-            trace!(
-                "No server status processor configured, logging server message: {}",
-                raw_message
-            );
-        }
-    }
-}
-
-impl PacketHandler for PacketRouter {
-    fn process_raw_message(&self, raw_message: &str) {
-        // Handle archival if configured
-        if let Some(archive) = &self.archive_service {
-            archive.archive(raw_message);
-        }
-    }
-
-    fn as_any(&self) -> &dyn std::any::Any {
-        self
-    }
-
-    fn process_packet(&self, packet: AprsPacket) {
-        match &packet.data {
-            AprsData::Position(_) => {
-                if let Some(pos_proc) = &self.position_processor {
-                    pos_proc.process_position_packet(&packet);
-                } else {
-                    trace!("No position processor configured, skipping position packet");
-                }
-            }
-            AprsData::Status(_) => {
-                trace!(
-                    "Received status packet from {} (source type: {:?})",
-                    packet.from,
-                    packet.position_source_type()
-                );
-                if let Some(status_proc) = &self.receiver_status_processor {
-                    status_proc.process_status_packet(&packet);
-                } else {
-                    warn!(
-                        "No receiver status processor configured, skipping status packet from {}",
-                        packet.from
-                    );
-                }
-            }
-            _ => {
-                debug!(
-                    "Received packet of type {:?}, no specific handler",
-                    std::mem::discriminant(&packet.data)
-                );
-            }
-        }
-    }
-}
-
-/// Processor for handling position packets from various sources
-pub struct PositionPacketProcessor {
-    /// Aircraft position processor for handling aircraft-specific logic
-    aircraft_processor: Option<AircraftPositionProcessor>,
-    /// Receiver position processor for handling receiver-specific logic
-    receiver_processor: Option<ReceiverPositionProcessor>,
-}
-
-impl Default for PositionPacketProcessor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl PositionPacketProcessor {
-    /// Create a new PositionPacketProcessor
-    pub fn new() -> Self {
-        Self {
-            aircraft_processor: None,
-            receiver_processor: None,
-        }
-    }
-
-    /// Add an aircraft position processor
-    pub fn with_aircraft_processor(mut self, processor: AircraftPositionProcessor) -> Self {
-        self.aircraft_processor = Some(processor);
-        self
-    }
-
-    /// Add a receiver position processor
-    pub fn with_receiver_processor(mut self, processor: ReceiverPositionProcessor) -> Self {
-        self.receiver_processor = Some(processor);
-        self
-    }
-
-    /// Process a position packet, routing based on source type
-    pub fn process_position_packet(&self, packet: &AprsPacket) {
-        match packet.position_source_type() {
-            PositionSourceType::Aircraft => {
-                if let Some(aircraft_proc) = &self.aircraft_processor {
-                    aircraft_proc.process_aircraft_position(packet);
-                } else {
-                    warn!("No aircraft processor configured, skipping aircraft position");
-                }
-            }
-            PositionSourceType::Receiver => {
-                if let Some(receiver_proc) = &self.receiver_processor {
-                    receiver_proc.process_receiver_position(packet);
-                } else {
-                    warn!(
-                        "No receiver processor configured, skipping receiver position from {}",
-                        packet.from
-                    );
-                }
-            }
-            PositionSourceType::WeatherStation => {
-                trace!(
-                    "Position from weather station {} - logging and ignoring",
-                    packet.from
-                );
-            }
-            source_type => {
-                trace!(
-                    "Position from unknown source type {:?} from {} - ignoring",
-                    source_type, packet.from
-                );
-            }
-        }
-    }
-}
-
-/// Processor for handling aircraft position packets
-pub struct AircraftPositionProcessor {
-    /// Fix processor for database storage
-    fix_processor: Option<FixProcessor>,
-    /// Flight detection processor for flight tracking
-    flight_detection_processor: Option<Arc<FlightTracker>>,
-}
-
-impl Default for AircraftPositionProcessor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl AircraftPositionProcessor {
-    /// Create a new AircraftPositionProcessor
-    pub fn new() -> Self {
-        Self {
-            fix_processor: None,
-            flight_detection_processor: None,
-        }
-    }
-
-    /// Add a fix processor for database storage
-    pub fn with_fix_processor(mut self, processor: FixProcessor) -> Self {
-        self.fix_processor = Some(processor);
-        self
-    }
-
-    /// Add a flight detection processor for flight tracking
-    pub fn with_flight_detection_processor(mut self, processor: Arc<FlightTracker>) -> Self {
-        self.flight_detection_processor = Some(processor);
-        self
-    }
-
-    /// Process an aircraft position packet
-    pub fn process_aircraft_position(&self, packet: &AprsPacket) {
-        let raw_message = packet.raw.clone().unwrap_or_default();
-
-        // Convert to Fix and process with fix processor if available
-        if let Some(ref fix_proc) = self.fix_processor {
-            fix_proc.process_aprs_packet(packet.clone(), &raw_message);
-        }
-
-        // Note: The flight detection processor is now handled inside FixProcessor
-        // so we don't need to call it separately here anymore
-    }
-}
-
-/// Processor for handling receiver position packets
-pub struct ReceiverPositionProcessor {
-    /// Repository for updating receiver locations
-    receiver_repo: ReceiverRepository,
-}
-
-impl ReceiverPositionProcessor {
-    /// Create a new ReceiverPositionProcessor
-    pub fn new(receiver_repo: ReceiverRepository) -> Self {
-        Self { receiver_repo }
-    }
-
-    /// Process a receiver position packet and update its location
-    pub fn process_receiver_position(&self, packet: &AprsPacket) {
-        // Extract position data from packet
-        if let AprsData::Position(position) = &packet.data {
-            let callsign = packet.from.to_string();
-            let latitude = position.latitude.as_();
-            let longitude = position.longitude.as_();
-
-            info!(
-                "Processing receiver position for {}: lat={}, lon={}",
-                callsign, latitude, longitude
-            );
-
-            // Update receiver position in database (async)
-            tokio::spawn({
-                let repo = self.receiver_repo.clone();
-                let callsign = callsign.clone();
-                async move {
-                    // First, get the receiver to obtain its ID
-                    match repo.get_receiver_by_callsign(&callsign).await {
-                        Ok(Some(receiver)) => {
-                            // Update position
-                            match repo
-                                .update_receiver_position(&callsign, latitude, longitude)
-                                .await
-                            {
-                                Ok(true) => {
-                                    debug!(
-                                        "Updated receiver position for {}: ({}, {})",
-                                        callsign, latitude, longitude
-                                    );
-
-                                    // Update latest_packet_at
-                                    if let Err(e) = repo.update_latest_packet_at(receiver.id).await
-                                    {
-                                        error!(
-                                            "Failed to update latest_packet_at for receiver {}: {}",
-                                            callsign, e
-                                        );
-                                    }
-                                }
-                                Ok(false) => {
-                                    warn!(
-                                        "Receiver {} not found in database, position not updated",
-                                        callsign
-                                    );
-                                }
-                                Err(e) => {
-                                    error!(
-                                        "Failed to update receiver position for {}: {}",
-                                        callsign, e
-                                    );
-                                }
-                            }
-                        }
-                        Ok(None) => {
-                            debug!(
-                                "Position packet from {} is not a known receiver, auto-inserting",
-                                callsign
-                            );
-
-                            // Auto-insert minimal receiver
-                            match repo.insert_minimal_receiver(&callsign).await {
-                                Ok(receiver_id) => {
-                                    info!(
-                                        "Auto-inserted receiver {} (id: {}), now updating position",
-                                        callsign, receiver_id
-                                    );
-
-                                    // Update position
-                                    match repo
-                                        .update_receiver_position(&callsign, latitude, longitude)
-                                        .await
-                                    {
-                                        Ok(true) => {
-                                            debug!(
-                                                "Updated receiver position for {}: ({}, {})",
-                                                callsign, latitude, longitude
-                                            );
-
-                                            // Update latest_packet_at
-                                            if let Err(e) =
-                                                repo.update_latest_packet_at(receiver_id).await
-                                            {
-                                                error!(
-                                                    "Failed to update latest_packet_at for receiver {}: {}",
-                                                    callsign, e
-                                                );
-                                            }
-                                        }
-                                        Ok(false) => {
-                                            warn!(
-                                                "Receiver {} not found in database after auto-insertion, position not updated",
-                                                callsign
-                                            );
-                                        }
-                                        Err(e) => {
-                                            error!(
-                                                "Failed to update receiver position for auto-discovered receiver {}: {}",
-                                                callsign, e
-                                            );
-                                        }
-                                    }
-                                }
-                                Err(e) => {
-                                    error!("Failed to auto-insert receiver {}: {}", callsign, e);
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            error!("Failed to look up receiver {}: {}", callsign, e);
-                        }
-                    }
-                }
-            });
-        } else {
-            warn!(
-                "Expected position packet from receiver {} but got different packet type",
-                packet.from
-            );
-        }
-    }
-}
-
-/// Processor for handling receiver status packets
-pub struct ReceiverStatusProcessor {
-    /// Repository for storing receiver status data
-    status_repo: ReceiverStatusRepository,
-    /// Repository for looking up receiver IDs by callsign
-    receiver_repo: ReceiverRepository,
-}
-
-impl ReceiverStatusProcessor {
-    /// Create a new ReceiverStatusProcessor
-    pub fn new(status_repo: ReceiverStatusRepository, receiver_repo: ReceiverRepository) -> Self {
-        Self {
-            status_repo,
-            receiver_repo,
-        }
-    }
-
-    /// Process a status packet from a receiver
-    pub fn process_status_packet(&self, packet: &AprsPacket) {
-        let source_type = packet.position_source_type();
-        let callsign = packet.from.to_string();
-
-        trace!(
-            "ReceiverStatusProcessor: Processing status packet from {} (source_type: {:?})",
-            callsign, source_type
-        );
-
-        // Process all status packets, not just those identified as receivers
-        // The receiver lookup will determine if it's actually a receiver
-        if let AprsData::Status(status) = &packet.data {
-            let status_comment = status.comment.clone();
-            let received_at = chrono::Utc::now();
-            let raw_data = packet.raw.clone().unwrap_or_default();
-
-            // Look up receiver ID asynchronously
-            tokio::spawn({
-                let receiver_repo = self.receiver_repo.clone();
-                let status_repo = self.status_repo.clone();
-                let raw_data = raw_data.clone();
-
-                async move {
-                    // Look up receiver by callsign
-                    match receiver_repo.get_receiver_by_callsign(&callsign).await {
-                        Ok(Some(receiver)) => {
-                            // Create NewReceiverStatus from status comment
-                            let new_status = NewReceiverStatus::from_status_comment(
-                                receiver.id,
-                                &status_comment,
-                                received_at, // packet timestamp
-                                received_at, // received_at
-                                raw_data.clone(),
-                            );
-
-                            // Insert receiver status
-                            match status_repo.insert(&new_status).await {
-                                Ok(_) => {
-                                    // Track receiver status update metric
-                                    counter!("receiver_status_updates_total").increment(1);
-
-                                    // Update receiver's latest_packet_at
-                                    if let Err(e) =
-                                        receiver_repo.update_latest_packet_at(receiver.id).await
-                                    {
-                                        error!(
-                                            "Failed to update latest_packet_at for receiver {}: {}",
-                                            callsign, e
-                                        );
-                                    }
-                                }
-                                Err(e) => {
-                                    error!(
-                                        "Failed to insert receiver status for {}: {}",
-                                        callsign, e
-                                    );
-                                }
-                            }
-                        }
-                        Ok(None) => {
-                            debug!(
-                                "Status packet from {} is not a known receiver, auto-inserting",
-                                callsign
-                            );
-
-                            // Auto-insert minimal receiver
-                            match receiver_repo.insert_minimal_receiver(&callsign).await {
-                                Ok(receiver_id) => {
-                                    info!(
-                                        "Auto-inserted receiver {} (id: {}), now inserting status",
-                                        callsign, receiver_id
-                                    );
-
-                                    // Create NewReceiverStatus from status comment
-                                    let new_status = NewReceiverStatus::from_status_comment(
-                                        receiver_id,
-                                        &status_comment,
-                                        received_at, // packet timestamp
-                                        received_at, // received_at
-                                        raw_data.clone(),
-                                    );
-
-                                    // Insert receiver status
-                                    match status_repo.insert(&new_status).await {
-                                        Ok(_) => {
-                                            // Track receiver status update metric
-                                            counter!("receiver_status_updates_total").increment(1);
-
-                                            // Update receiver's latest_packet_at
-                                            if let Err(e) = receiver_repo
-                                                .update_latest_packet_at(receiver_id)
-                                                .await
-                                            {
-                                                error!(
-                                                    "Failed to update latest_packet_at for receiver {}: {}",
-                                                    callsign, e
-                                                );
-                                            }
-                                        }
-                                        Err(e) => {
-                                            error!(
-                                                "Failed to insert receiver status for auto-discovered receiver {}: {}",
-                                                callsign, e
-                                            );
-                                        }
-                                    }
-                                }
-                                Err(e) => {
-                                    error!("Failed to auto-insert receiver {}: {}", callsign, e);
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            error!("Failed to look up receiver by callsign {}: {}", callsign, e);
-                        }
-                    }
-                }
-            });
-        } else {
-            warn!(
-                "Expected status packet but got different packet type from {}",
-                callsign
-            );
-        }
-    }
-}
-
-/// Processor for handling APRS server comment/status messages
-pub struct ServerStatusProcessor {
-    /// Repository for storing server messages
-    server_messages_repo: ServerMessagesRepository,
-}
-
-impl ServerStatusProcessor {
-    /// Create a new ServerStatusProcessor
-    pub fn new(server_messages_repo: ServerMessagesRepository) -> Self {
-        Self {
-            server_messages_repo,
-        }
-    }
-
-    /// Process a server status message (line starting with #)
-    pub async fn process_server_message(&self, raw_message: &str) {
-        let received_at = Utc::now();
-
-        // Parse server message format: # aprsc 2.1.15-gc67551b 22 Sep 2025 21:51:55 GMT GLIDERN1 51.178.19.212:10152
-        if let Some(parsed) = self.parse_server_message(raw_message, received_at) {
-            if let Err(e) = self.server_messages_repo.insert(&parsed).await {
-                debug!("Failed to insert server message: {}", e);
-            } else {
-                trace!(
-                    "Successfully stored server message from {}",
-                    parsed.server_name
-                );
-            }
-        } else {
-            debug!("Failed to parse server message: {}", raw_message);
-        }
-    }
-
-    /// Parse server message string into ServerMessage
-    /// Expected format: # aprsc 2.1.15-gc67551b 22 Sep 2025 21:51:55 GMT GLIDERN1 51.178.19.212:10152
-    fn parse_server_message(
-        &self,
-        raw_message: &str,
-        received_at: DateTime<Utc>,
-    ) -> Option<ServerMessage> {
-        let trimmed = raw_message.trim_start_matches('#').trim();
-        let parts: Vec<&str> = trimmed.split_whitespace().collect();
-
-        trace!("Parsing server message parts: {:?}", parts);
-
-        // Need at least: software version, date (3 parts), time, GMT, server_name, endpoint
-        if parts.len() < 8 {
-            debug!(
-                "Server message has {} parts, need at least 8: {}",
-                parts.len(),
-                raw_message
-            );
-            return None;
-        }
-
-        // Extract software (first two parts: "aprsc 2.1.15-gc67551b")
-        let software = format!("{} {}", parts[0], parts[1]);
-
-        // Extract timestamp parts: "22 Sep 2025 21:51:55 GMT"
-        let date_time_str = format!(
-            "{} {} {} {} {}",
-            parts[2], parts[3], parts[4], parts[5], parts[6]
-        );
-
-        // Parse timestamp
-        let server_timestamp =
-            match NaiveDateTime::parse_from_str(&date_time_str, "%d %b %Y %H:%M:%S GMT") {
-                Ok(naive_dt) => naive_dt.and_utc(),
-                Err(_) => {
-                    debug!("Failed to parse timestamp: {}", date_time_str);
-                    return None;
-                }
-            };
-
-        // Extract server name and endpoint
-        let server_name = parts[7].to_string();
-        let server_endpoint = parts[8].to_string();
-
-        Some(ServerMessage::new(
-            software,
-            server_timestamp,
-            received_at,
-            server_name,
-            server_endpoint,
-        ))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use chrono::NaiveDateTime;
     use std::sync::atomic::{AtomicUsize, Ordering};
 
     #[test]
